moving new code to separate files to avoid adding large amounts of emscripten-only internals to main.cpp
